+++ packages/main/src/components/ActionSheet/ActionSheet.jss.ts
@@ -5,8 +5,9 @@ const styles = {
    *  is being applied to the encapsulating ul element
    */
   actionSheet: {
-    margin: 0,
-    padding: '0.1875rem 0.375rem',
+    '&::part(content)': {
+      padding: '0.1875rem 0.375rem'
+    },
     '& [ui5-button]': {
       display: 'block'
     },