+++ packages/main/src/components/Toolbar/OverflowPopover.tsx
@@ -5,12 +5,12 @@ import { ButtonDesign } from '../../enums/ButtonDesign';
 import { PopoverPlacementType } from '../../enums/PopoverPlacementType';
 import { stopPropagation } from '../../internal/stopPropagation';
 import { Popover, PopoverDomRef } from '../../webComponents/Popover';
-import { ToggleButton } from '../../webComponents/ToggleButton';
+import { ToggleButton, ToggleButtonDomRef } from '../../webComponents/ToggleButton';
 import { useSyncRef } from '@ui5/webcomponents-react-base';
 
 interface OverflowPopoverProps {
   lastVisibleIndex: number;
-  contentClass: string;
+  classes: any;
   children: ReactNode;
   portalContainer: Element;
   overflowContentRef: Ref<HTMLDivElement>;
@@ -22,7 +22,7 @@ interface OverflowPopoverProps {
 export const OverflowPopover: FC<OverflowPopoverProps> = (props: OverflowPopoverProps) => {
   const {
     lastVisibleIndex,
-    contentClass,
+    classes,
     children,
     portalContainer,
     overflowContentRef,
@@ -33,10 +33,12 @@ export const OverflowPopover: FC<OverflowPopoverProps> = (props: OverflowPopover
 
   const [componentRef, popoverRef] = useSyncRef<PopoverDomRef>(overflowPopoverRef);
   const [pressed, setPressed] = useState(false);
+  const toggleBtnRef = useRef<ToggleButtonDomRef>(null);
 
   const handleToggleButtonClick = useCallback(
     (e) => {
       e.stopPropagation();
+      toggleBtnRef.current.accessibilityAttributes.expanded = `${!pressed}`;
       if (popoverRef.current) {
         if (!pressed) {
           popoverRef.current.showAt(e.target);
@@ -85,9 +87,19 @@ export const OverflowPopover: FC<OverflowPopoverProps> = (props: OverflowPopover
     });
   }, [children, lastVisibleIndex]);
 
+  useEffect(() => {
+    customElements.whenDefined('ui5-toggle-button').then(() => {
+      if (toggleBtnRef.current) {
+        toggleBtnRef.current.accessibilityAttributes.expanded = `${pressed}`;
+        toggleBtnRef.current.accessibilityAttributes.hasPopup = 'menu';
+      }
+    });
+  }, []);
+
   return (
     <>
       <ToggleButton
+        ref={toggleBtnRef}
         design={ButtonDesign.Transparent}
         icon="overflow"
         onClick={handleToggleButtonClick}
@@ -95,8 +107,14 @@ export const OverflowPopover: FC<OverflowPopoverProps> = (props: OverflowPopover
         accessibleName={showMoreText}
       />
       {createPortal(
-        <Popover placementType={PopoverPlacementType.Bottom} ref={componentRef} onAfterClose={handleClose} hideArrow>
-          <div className={contentClass} ref={overflowContentRef}>
+        <Popover
+          className={classes.popover}
+          placementType={PopoverPlacementType.Bottom}
+          ref={componentRef}
+          onAfterClose={handleClose}
+          hideArrow
+        >
+          <div className={classes.popoverContent} ref={overflowContentRef}>
             {renderChildren()}
           </div>
         </Popover>,